### 2019-08-11 #1640 @whedon
[PRE REVIEW]: pySRURGS - a python package for symbolic regression by uniform random global search
**Submitting author:** @pySRURGS (<a href="http://orcid.org/0000-0002-3050-8943">Sohrab Towfighi</a>)
**Repository:** <a href="https://github.com/pySRURGS/pySRURGS" target ="_blank">https://github.com/pySRURGS/pySRURGS</a>
**Version:** 0.1
