### 2019-08-11 5d37c53
Update paper.md
Former-commit-id: 6514529cd196c9a809c744dd8b2e22248d3aac7b
