+++ paper/paper.md
@@ -23,13 +23,13 @@ Symbolic regression is a regression analysis where we search for the
 mathematical expression that best describes the relationship between our 
 independent variables and our output variable. ``pySRURGS`` is a python code 
 for symbolic regression by uniform random global search. Typically, symbolic
-regression problems are solved using genetic programming[@Koza:1992], but the 
+regression problems are solved using genetic programming [@Koza:1992], but the 
 data demonstrating that genetic programming actually outperforms random search 
 on symbolic regression problems is entirely lacking. Comparing the performance 
 of an algorithm against that of random search demonstrates the advantage gained 
 by the algorithm's internal machinery. Our review of the literature found a 
 large study examining the parameter space of genetic programming, which found 
-few patterns with regard to how users should select genetic algorithm 
+few patterns with regards to how users should select genetic algorithm 
 hyperparameters [@Sipper:2018]. These workers attempt to compare their genetic 
 algoritm's performance to that of random search, but we reproduce their work 
 and find they unfairly disadvantaged the random search by forcing it to only 
