### 2019-08-11 39c7d61
compile_results printout
Former-commit-id: 87567b976dfd629fa47c0afeb02b730bad4f63a3
