+++ README.md
@@ -101,16 +101,16 @@ A sample problem is provided. The filename denotes the true equation.
 
 ```
 
-$ winpty python pySRURGS.py -max_num_fit_params 5 -max_permitted_trees 1000 ./csvs/x1_squared_minus_five_x3.csv 10
-Running in single processor mode
-100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:27<00:00,  2.74s/it]
-  Normalized Mean Squared Error        R^2  Equation, simplified                                                                                           Parameters
--------------------------------  ---------  -------------------------------------------------------------------------------------------------------------  ----------------------------------------------
-                        10.2672   0.963167  -(x0 - x1*((p1*p3 - x1*x3)/(p1*x1))**x3*(p3 - x0))*(p0*x1*(x0*(x1 + x3) - x1) - x0*sin(p2))/(x0*x1*(p3 - x0))  -4.74E-04,1.51E+04,2.14E-01,5.25E+03,-1.27E+04
-                        17.338    0.941908  x2 + p4*x3/p1**2 - p0**(-x1)*p4*x5*log(p0)/x4 - p0**(-x1)*x2*x5*log(p0)/x4                                     1.00E+00,1.69E+02,1.00E+00,1.00E+00,1.33E+05
-                        65.4971   0.742457  (2*p3 - p4**x2)*(-p0*p3*x0 + p0*x2 + x0*x2)*((-p2 + p3)**(x3/p0))**(-p1 + x2)/(p0*x0*sin(p1))                  1.88E+00,-1.15E+00,-3.12E+00,4.76E-01,1.07E-01
-                       111.863    0.461941  x5**(-x4)*(-x3**x5*(p4**p0 - 2*x5))**sinh(p0)*(p0*x5 - p1*x0)/(p1*(p2 - x1))                                   8.75E-01,5.85E-04,1.00E+00,1.00E+00,5.76E+00
-                       256.675   -1.05228   (x1 + x2)*(x2 + x4)*(p0 - p0**p1 + x2)*(p0 - x0 + x3 + x5)*log(p0)                                             1.03E+00,1.03E+00,1.00E+00,1.00E+00,1.00E+00
+$ winpty python pySRURGS.py -max_num_fit_params 0 -max_permitted_trees 1000 ./csvs/quartic_polynomial.csv 100
+Running in multi processor mode
+1008it [01:13, 12.48it/s]
+  Normalized Mean Squared Error       R^2  Equation, simplified                       Parameters
+-------------------------------  --------  -----------------------------------------  ------------
+                       0.21425   0.995187  x*(x*(2*x + 1) + 2)
+                       0.22746   0.994465  x*(4*x - (x + 1)**x + 1)
+                       0.289212  0.992846  x*(x + 1)**(x + 1) + x**(x + 2)*(x**x)**x
+                       0.363446  0.991796  ((2*x)**x*(x + 1))**(x**(2*x))
+                       0.41818   0.987663  0**x + (2*x)**x*(x + 1)**x
 ```
 
 ### Important details 
