### 2019-08-11 258eab5
Update evaluate_SRGP_on_benchmarks.py
decreased the population size for feasibility

Former-commit-id: ea4699ea88aa9ae6da09a8c3ea7bffb2505dd702 [formerly 3789cdbc1f9e516ed8da2f6bd52112784249e895]
Former-commit-id: ad55d4af2f4b096c0d51309f28a2106c186ae423
Former-commit-id: 61ffa04b2b99dd23e85b78d478926aebef9546ac
