### 2019-08-11 fb84549
Create .travis.yml
Former-commit-id: 93d5d764366190f0cc67a36ca6bf3a3bb6e2960f [formerly 1cc070966350b21fca2656a398ab8676763f3711]
Former-commit-id: 02d85552a5253657db76fba8dbd1674d15ce9853
Former-commit-id: 015e6fa63368bfd2a819f55da7327983c39b4e52
