+++ .travis.yml
@@ -1,0 +1,11 @@
+language: python
+python:
+  - "3.6"
+install: # command to install dependencies
+  - "pip install -r requirements.txt"
+# command to run tests
+script: python test.py
+
+notifications:
+  email:
+    - sohrab.towfighi@mail.utoronto.ca
