### 2019-08-11 daeedb3
Update pySRURGS.py
removed seed(0)

Former-commit-id: b7a5ba2001726b32f435fbe42879ededd20246c5 [formerly d565b4c34c1e3f05ebd8caba0273f586d93b6263]
Former-commit-id: 3c8c3260b08ce9d49eaa2ea1bab0e93ff4a7d984
Former-commit-id: e9b98b4485a8bc4f4d4d3f8c39258e0da0fc4336
