### 2019-08-11 260aecd
install sh in the testing environment
Former-commit-id: c55db03904d8e67b8ae9a389ef37e9c3df5721ce [formerly 05e107fbb1b3d54bc7dc21edace80ff3a46f2b9c]
Former-commit-id: 3465fd001356a3db028fc4d4e1d60e18c65e843c
Former-commit-id: d5fa71420c1ed2fdea8d02247b2058c2feafa1d0
