+++ .travis.yml
@@ -3,6 +3,7 @@ python:
   - "3.6"
 install: # command to install dependencies
   - "pip install -r requirements.txt"
+  - "pip install sh"
 # command to run tests
 script: python test.py
 
