+++ test.py
@@ -1,0 +1,6 @@
+import sh
+import os
+import sys 
+
+sh.python('pySRURGS.py', './csvs/quartic_polynomial.csv', 100)
+
