### 2019-08-11 b00dece
added more tests, still using sh
Former-commit-id: 91e96defc77b06aa54845a4d2cded6312cc236cb
