### 2019-08-11 4610f57
fewer evals for the test suite
Former-commit-id: 508af0f5845f4021e0fb08807b054537c8cee25a [formerly 8b0c3c029a7a56a6ed9dc9975988c08460e38d84]
Former-commit-id: 61d7689e9fd6e45704f17b072002134527a77307
Former-commit-id: 758cddc779b104fd20751451d8368ffaca3e727d
