### 2019-08-11 2ea0238
revised number of fit params error
'You have more fitting parameters than you do data points' -> this exception was incorrectly implemented.

Former-commit-id: 003acceb16cf178fd5e75819e55a77f694b0ccda [formerly 1df8a2c1b8fc1712c2f633f71a072cf2bae80932]
Former-commit-id: e555a851a3059344d42c30e466fa0354d8ef80fb
Former-commit-id: 47c9192890030ac2dce0cd56510daf560b036cdf
