+++ test.py
@@ -6,8 +6,8 @@ sh.python('pySRURGS.py', './csvs/quartic_polynomial.csv', 10)
 sh.python('pySRURGS.py', '-single', './csvs/quartic_polynomial.csv', 10)
 sh.python('pySRURGS.py', '-count', './csvs/quartic_polynomial.csv', 10)
 sh.python('pySRURGS.py', '-benchmarks', './csvs/quartic_polynomial.csv', 10)
-sh.python('pySRURGS.py', '-num_max_fit_params', 0, './csvs/quartic_polynomial.csv', 10)
-sh.python('pySRURGS.py', '-num_max_fit_params', 5, './csvs/quartic_polynomial.csv', 10)
+sh.python('pySRURGS.py', '-max_num_fit_params', 0, './csvs/quartic_polynomial.csv', 10)
+sh.python('pySRURGS.py', '-max_num_fit_params', 5, './csvs/quartic_polynomial.csv', 10)
 sh.python('pySRURGS.py', '-funcs_arity_two', 'add,sub,div', '-num_max_fit_params', 5, './csvs/quartic_polynomial.csv', 10)
 sh.python('pySRURGS.py', '-funcs_arity_one', 'tan,exp', '-num_max_fit_params', 5, './csvs/quartic_polynomial.csv', 10)
 sh.python('pySRURGS.py', '-max_permitted_trees', 10, '-num_max_fit_params', 5, './csvs/quartic_polynomial.csv', 10)
