### 2019-08-12 2b40cae
gpl3 badge
Former-commit-id: 270abf033a3852b653830db87da1718a8bcd0235
