### 2019-08-12 913280a
python badge
Former-commit-id: 2b55bbbd4aa614a6bf85e0018b226e46111c364c [formerly 30a2388e899916ee22936b9edd36bbab7815a927]
Former-commit-id: c40c4c2821ea66ebfc244be09beffc7d7d7fbdf0
Former-commit-id: 558382ce5250fb0e1357dc386345137c4ef8ce56
