### 2019-08-12 99270df
Update README.md
cleaned up the features section
