### 2019-08-12 6c8565a
removed older experiment results and updating script to distribute tasks
