+++ experiments/database.py
@@ -10,6 +10,7 @@ from secret import *
 import pymysql
 import sshtunnel
 import platform 
+import argparse 
 
 try:
     import sh
@@ -80,7 +81,7 @@ def purge_db():
                                              port=tunnel.local_bind_port,
                                              database=DATABASE_NAME)
         mycursor = mydb.cursor()                               
-        sql = "DROP TABLE jobs'"
+        sql = "DROP TABLE jobs"
         mycursor.execute(sql)
         mydb.commit()
         mydb.close()
@@ -159,4 +160,20 @@ def run_all_SRGP_jobs():
             i = i + 1
 
 if __name__ == '__main__':
-    run_all_SRGP_jobs()
+    # Read the doc string at the top of this script.
+    # Run this script in terminal with '-h' as an argument.
+    parser = argparse.ArgumentParser(prog='database.py', formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+    parser.add_argument("-run_SRGP", help="run the code against all the SRGP problems in the mySQL database", action="store_true")
+    #parser.add_argument("run_SRURGS", help="run the code against all the SRURGS problems in the mySQL database")
+    parser.add_argument("-purge_db", help="deletes all the jobs in the database", action="store_true")
+    if len(sys.argv) < 2:
+        parser.print_usage()
+        sys.exit(1)
+    arguments = parser.parse_args()
+    if arguments.run_SRGP and arguments.purge_db:
+        raise Exception("Cannot do both run SRGP jobs and purge database")
+    if arguments.run_SRGP:
+        run_all_SRGP_jobs()
+    if arguments.purge_db:
+        purge_db()
+    
