### 2019-08-12 7da2382
error handling and submit to dropbox
Former-commit-id: 873877e44a0625f4aa0eca7c4c12d4f207756239 [formerly d9a1ba53ea5836ca7f3f57b264b092178981eee8]
Former-commit-id: ef2d6b7fb39370422690ee895912f0b01c70a052
Former-commit-id: 86e0ae138c4391b8afa969f3179c71196878efea
