### 2019-08-12 9860327
more manageable computations
Former-commit-id: 28d2eb4d9a82f9ecb87dff6c9d18c13c19c4f3ca [formerly 6c26606b4ecba3e7433be31f13d50d87a0a0d856]
Former-commit-id: 83acc65f420fd8bb6d213c095afc76d886a7cb6e
Former-commit-id: 18a889d01f2176de52b31a7a52d61eddb3c6a391
