+++ experiments/submit_SRGP_on_benchmark_jobs_to_db.py
@@ -38,7 +38,7 @@ def generate_list_of_experiments(SR_config, start_index, count_experiments, n_ru
     funcs_arity_one = ','.join(SR_config._f_functions)
     max_num_fit_params = SR_config._max_num_fit_params
     max_size_trees = SR_config._max_permitted_trees
-    popsize = 300
+    popsize = 100
     numgens = 20
     # first twenty problems
     for z in range(start_index,count_experiments):
