### 2019-08-14 b1347fb
significant edits to SRGP after found issues in string representatino of equations. Also, varAnd_EnsureValid has been parallelized
Former-commit-id: 51721da21d7103a555aeaa4dfa7da74fda95145e [formerly 52df8ec46d6af942b526fe228865588f0acf7dc8]
Former-commit-id: 370887aececa731ebabdba491c735e782d02fed8
Former-commit-id: 7a271ff1960bf798bb30db95888d7d7775ae76bc
