+++ experiments/SRGP.py
@@ -1,6 +1,9 @@
 # Adapting https://github.com/usnistgov/pysr
 # to the pySRURGS code style and datastructure
 # Some improvements to performance through checking previously considered equations
+
+# TODO, remove the variable and parameter tags simplification.
+
 import warnings
 warnings.simplefilter('ignore')
 import scipy
@@ -25,8 +28,13 @@ import pandas as pd
 sys.path.append('./..')
 
 # making the codes analogous 
-from pySRURGS import add, mul, sub, div, pow, sin, cos, tan, log, exp, sinh, tanh, cosh
-from pySRURGS import Dataset, SymbolicRegressionConfig, simplify_equation_string, eval_equation, create_fitting_parameters, initialize_db, Result, check_goodness_of_fit
+from pySRURGS import (add, mul, sub, div, pow, sin, cos, tan,
+                      log, exp, sinh, tanh, cosh,
+                      Dataset, SymbolicRegressionConfig, eval_equation,
+                      simplify_equation_string, create_fitting_parameters,
+                      create_parameter_list, create_variable_list,
+                      initialize_db, check_goodness_of_fit, Result,
+                      clean_funcstring, remove_dict_tags)
 
 # load the arguments
 parser = argparse.ArgumentParser(prog='pySR.py', formatter_class=argparse.ArgumentDefaultsHelpFormatter)
@@ -92,20 +100,20 @@ if 'cosh' in f_functions:
 if 'tanh' in f_functions:
     pset.addPrimitive(tanh, 1, name='tanh')
 
-# dynamically create DEAP variables based on the 
-# number of variables in our raw data
-for i in range(0,len(dataset._x_labels)):     
-    var_label = "df['" + dataset._x_labels[i] + "']"
+# dynamically create DEAP variables based on the var names and parameter names 
+var_names = create_variable_list(csv_path)
+for i in range(0,len(var_names)):     
+    var_label = var_names[i]
     deap_var_name = "ARG" + str(i)
     var_assignment_command = 'pset.renameArguments(' 
     var_assignment_command += deap_var_name + '="' + var_label + '")'
     eval(var_assignment_command)
 
-for int_param in range(0, int_max_params):# dynamically create DEAP 
-    var_label = "pd['p" + str(int_param) + "']"
-    deap_var_name = "ARG" + str(len(dataset._x_labels)+int_param)
+param_names = create_parameter_list(int_max_params)
+for i in range(0, int_max_params):
+    deap_var_name = "ARG" + str(len(var_names) + i)
     var_assignment_command = 'pset.renameArguments(' 
-    var_assignment_command += deap_var_name + '="' + var_label + '")'
+    var_assignment_command += deap_var_name + '="' + param_names[i] + '")'
     eval(var_assignment_command)
 
 #Define our 'Individual' class
@@ -116,40 +124,32 @@ creator.create('Individual',
                fitness=creator.FitnessMin,
                params=create_fitting_parameters(int_max_params))               
 
+def strip_dictionary_tags(dataset, equation_string):
+    terminals_list = dataset._terminals_list
+    for terminal in terminals_list:
+        equation_string = equation_string.replace("df['"+terminal+"']", terminal)
+        equation_string = equation_string.replace("pd['"+terminal+"']", terminal)
+    return equation_string
+
 #Define our evaluation function
-
-def stringify(individual):
-    """Return the expression in a human readable string.
-    """
-    string = ""
-    if type(individual) == list:
-        individual = individual[0]
-    stack = []
-    for node in individual:
-        stack.append((node, []))
-        while len(stack[-1][1]) == stack[-1][0].arity:
-            prim, args = stack.pop()
-            string = prim.format(*args)
-            if len(stack) == 0:
[…]
