### 2019-08-14 51721da
significant edits to SRGP after found issues in string representatino of equations. Also, varAnd_EnsureValid has been parallelized
Former-commit-id: 52df8ec46d6af942b526fe228865588f0acf7dc8
