### 2019-08-14 0d2669c
removed duplicate file
Former-commit-id: 7bbd3a0942bba744e94740af3a0263ee12aaf962
