### 2019-08-14 69ecd0b
removing scoop dependence
Former-commit-id: 9e293eec9d04116a88c27d5bc0d061d4778b336d [formerly 96b3b7e1a4cb509ad164f586818228515859384a]
Former-commit-id: 25e55cc49d14a464f548bae974d1188fc21fdc97
Former-commit-id: 7861291591548a1129b960604895ee08d1a7ddb7
