### 2019-08-15 b68c758
redid the benchmarks preventing y_data be a constant
Former-commit-id: cd7d4863a851be17c8cbde162b914629e6ffc5c3
