+++ csvs/benchmarks/1_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 3.84011707  3.70683643  3.33052542  0.46213038 -6.01644963]
+Fitting parameters: [ 9.15021409 -6.50436822 -0.53319776  9.4234529   7.4621563 ]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: 0**x2*(x1*(x0 - x2))**p0 + (x1 + x4)**p0 + (-p0 + x0 + x4)**x1
-Raw equation: pow(sub(add(add(add(x4, x0), p0), x1), pow(add(x4, x1), p0)), pow(mul(mul(x1, sub(x0, x2)), p0), pow(sub(x0, x0), x2)))
+Simplified equation: (p1*(p1 + x0**x2*x1 - x1) + x4)*(x0**x2*x3 - x4*(p0 + p2)**x2)/(p1*x4)
+Raw equation: sub(add(mul(div(add(x4, p1), p1), x1), mul(pow(x0, x2), x1)), sub(mul(div(x3, pow(x0, x2)), x4), pow(add(p0, p2), x2)))
