+++ csvs/benchmarks/80_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-8.92785307  0.44004556 -3.68430107  0.89162494  3.40896642]
+Fitting parameters: [ 8.75330847 -7.22837027 -4.28542905  3.90087431 -6.14471886]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (((x3 + x4**2)/x4)**x3 + (p1*x0 + x1)*(x0*x1 + x3 + (-p0 + x1 + x2)**x3))/(p1*x0 + x1)
-Raw equation: div(div(add(add(pow(x3, x4), x4), x3), add(mul(x0, p1), x1)), pow(add(sub(x1, sub(p0, x2)), x3), mul(add(x1, x0), x3)))
+Simplified equation: p0*p1*x0*x2*(x4**(-x4)*(p0*x4**x4 + x3))**((p1*x2 + x1)/p1)*(-p1 + x3)**p2
+Raw equation: mul(mul(mul(mul(mul(x0, p1), x2), p0), pow(sub(x3, p1), p2)), div(pow(add(x3, pow(x4, x4)), p0), add(div(x1, p1), x2)))
