+++ csvs/benchmarks/91_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-4.87575923  7.64250242 -5.38196558  2.90622935  6.94026553]
+Fitting parameters: [ 0.57852191 -8.19431    -0.31158913  5.97086432  0.52257833]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (p2 - x2**x1)*(p0*(x2 - x3) - p1)*(x0*(x4/x0)**x4 + x3**(p2 - x2))/(p0*x0)
-Raw equation: div(mul(mul(sub(add(p1, p0), x3), x2), sub(pow(x2, x1), p2)), div(add(pow(x3, sub(p2, x2)), x0), div(pow(x4, x0), x4)))
+Simplified equation: -(p0 + (x2 + x3)*(p1*x0 - p1 + x1))*(p2*(p2 + x3) + x1*x4*(p2*x0 - x4))/(p2*x4*(x2 + x3))
+Raw equation: div(mul(sub(mul(sub(x4, p2), x0), x1), add(div(p2, x3), x4)), add(div(add(p0, add(x2, x3)), x1), sub(mul(x0, p1), p1)))
