+++ csvs/benchmarks/92_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 7.27274014  0.93725638  5.74508643 -9.01335019 -3.05293196]
+Fitting parameters: [ 8.1399993   1.51103707 -1.67417379 -2.98570292 -4.66004207]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p2**(-p1 - x2)*x4*(x1 + x3*x4)*(x2**x1)**x0/(p0*x0*x3*(x1 + x2))
-Raw equation: div(div(div(mul(add(x1, x3), x4), x4), add(mul(x1, x2), p0)), div(pow(mul(p2, add(p1, x2)), x0), pow(pow(x2, x1), x0)))
+Simplified equation: -p0 + p0/(x0*x4) + p1 + x1 + x2 - x3 + x3/x4 + 2*x4 - x1/p1
+Raw equation: add(add(div(add(sub(x3, x4), x1), p0), add(add(p1, x2), x4)), sub(div(sub(p0, mul(x4, x0)), x3), sub(div(x1, p1), x4)))
