+++ csvs/benchmarks/16_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 8.15527812 -0.09306441 -6.7780368  -5.80543608  4.79218798]
+Fitting parameters: [ 2.66046359 -7.09575855  0.03927121  0.69399946 -0.53654981]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: x0*x1*x3*(p0 + 1)*(p0 - x1 - x3 + x4)/(p0*p1*p2*x2)
-Raw equation: div(add(mul(mul(div(x0, p0), x0), x3), mul(mul(x3, p1), x2)), sub(mul(add(p0, sub(x4, x3)), x1), mul(div(x3, p2), x1)))
+Simplified equation: p1*x2*(p0*(-x1 + x4) + x1)*(p2*x1*(p0 + x3) + x2**(p1 - p2))/(p0*x1*x3)
+Raw equation: div(sub(mul(add(mul(x1, p0), x4), x1), mul(div(p1, x2), x3)), add(div(pow(x2, sub(p1, p2)), x1), mul(add(x3, p0), p2)))
