+++ csvs/benchmarks/31_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 8.54394143  3.66825569 -3.70183201  7.85904095  3.34149978]
+Fitting parameters: [-1.25465262  0.46350224  0.2465812   9.35702015  0.05865215]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (p0*(p1/x0)**x0*((p0 - x0*x1)/x1)**x0 - x0*(x0 + x3)*(p0**x2)**p2)/p0
-Raw equation: mul(sub(sub(div(pow(p0, x1), x0), x0), pow(div(p1, x0), x0)), mul(div(mul(x0, add(x3, x0)), p0), pow(pow(p0, x2), p2)))
+Simplified equation: p1*p2*x4*(p0 + x0*(p2 + x0))*(p0**3 + ((x1*x3 + x4)/x3)**x3)
+Raw equation: add(mul(add(pow(div(x4, x3), x1), x3), mul(mul(p0, p0), p0)), mul(add(mul(x0, add(x0, p2)), p0), mul(mul(p2, p1), x4)))

[…]
