### 2019-08-15 e8c019e
redid the benchmarks preventing y_data be a constant
Former-commit-id: b68c7580a7810c0600249c96260a38250e462b08 [formerly cd7d4863a851be17c8cbde162b914629e6ffc5c3]
Former-commit-id: dbf79b565d0fc3465e6dbdd56d17575e7a080f9c
Former-commit-id: 275c0739bf744e67aab694665cbc2b08506c1354
