+++ csvs/benchmarks/21_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-4.21520326 -4.39846823 -4.60061278  9.60314303 -4.78349959]
+Fitting parameters: [-9.05798669  5.58659428  4.13833962 -5.40406418 -0.30901031]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: ((p0*(-p0 + x2) + x2)/p0)**(x3**x4/x4)*(-p1 - x1 - x2**p2 + x4)/x3
-Raw equation: add(sub(div(mul(pow(x2, p0), x2), p0), div(pow(x3, x4), x4)), div(sub(sub(x4, add(x1, p1)), x3), div(pow(x2, p2), x3)))
+Simplified equation: (-p0*x4*(p1 + x3*(p1 - x2))*(x1 + 1) + x3*(p1 + p2*x0)**x1*(p0*p1 - p2))/(x3*x4*(x1 + 1)*(p0*p1 - p2))
+Raw equation: sub(add(add(div(div(p1, x3), p1), x2), sub(div(p2, p0), p1)), pow(div(add(p1, mul(x0, p2)), x1), add(mul(x1, x4), x4)))
