+++ csvs/benchmarks/32_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 8.09857451  4.3368033  -6.47875674  0.60313187  5.3823665 ]
+Fitting parameters: [ 5.54055432  9.58189669  2.94117364  9.48281322 -8.83676168]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: ((p0**2 + p1)/p0)**p1*(p1*(x0 + x1*x4**(p0 + x0)) + p2)/(p1*x3*(p0 - p1))
-Raw equation: div(mul(div(pow(add(p1, p0), p0), p1), mul(sub(p0, p1), x3)), mul(pow(add(x4, add(p0, x0)), x1), add(div(p2, p1), x0)))
+Simplified equation: -p0 - p1*p2*x2 + x0*(p0 - x1 + x4) + x1**x0 - x1**(-p0 + p1)*x3
+Raw equation: sub(sub(add(add(mul(p0, x4), x1), x0), sub(pow(x1, x0), p0)), add(pow(mul(x1, sub(p1, p0)), x3), mul(mul(x2, p2), p1)))
