+++ csvs/benchmarks/8_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-5.54799289 -6.5580132  -9.99639052  5.25723456  4.8117646 ]
+Fitting parameters: [ 2.98243223 -4.23159031 -8.5398977   9.3009809   0.01245517]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p1*x3 - p2 + x0*x3 + x0 + x3 + (x1 + x2)**x3 - x4/p2
-Raw equation: add(add(pow(add(add(x1, x2), x3), x0), mul(add(p1, x3), x3)), sub(sub(mul(x3, sub(x0, x0)), x0), add(div(x4, p2), p2)))
+Simplified equation: (((x1 + x3*(p1 - x4))/x3)**x3)**((-x3*(p2 + x0*x2) + 1)/x3)
+Raw equation: sub(add(pow(pow(div(x1, x3), p1), x4), add(sub(x3, p1), p1)), pow(div(sub(x3, sub(x0, x0)), x3), add(mul(x0, x2), p2)))
