+++ csvs/benchmarks/11_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 4.78265082  6.93531287  6.51516273 -8.44955757  9.61598757]
+Fitting parameters: [ 3.35188946 -2.93118257 -3.39666344  4.71263333  2.48718486]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (x1**2 - x2)*(p1*x0**2 - x1*x4)*(-p0 + p2 + x3)**x3/(p1*p2*x0*x1**2)
-Raw equation: add(mul(sub(mul(pow(p2, x3), p0), x3), div(sub(x2, x1), x1)), sub(div(div(x4, mul(p1, x0)), p2), div(div(x0, x1), p2)))
+Simplified equation: p2*(x2 - 1)*(x0 - x1 + x2)/(p1 + x0 - x3 + x4)
+Raw equation: mul(sub(mul(div(div(p2, x2), p2), x2), sub(add(x0, x2), x1)), sub(pow(mul(p0, sub(x2, x2)), x3), add(add(x0, p1), x4)))

[…]
