### 2019-08-15 3be6555
redid the benchmarks preventing y_data be a constant
Former-commit-id: f2d7ec715b833e7b6d93a6272f6b599bbbeb109f [formerly 06ff1833d40ea04c7143f8397212cbcce95d6757]
Former-commit-id: be36a9339875c2b93a23f6b285127b01759b27aa
Former-commit-id: 4a38712f4ec92441620b858bf315005e066b0cc4
