+++ csvs/benchmarks/89_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-0.57773019  0.68599077  2.59955351  2.88499039  6.77019088]
+Fitting parameters: [ 3.13244558 -0.73669943  4.28804173  5.40197175 -8.54803403]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: -x4*(-(p1*p2 - x3)/p2)**x2/(p2*(p0*p1**2 + x4*(p2 - x1**p1)))
-Raw equation: div(div(sub(div(pow(x3, p2), p1), x2), sub(sub(x1, p2), x1)), sub(mul(mul(p0, div(p1, x4)), p1), sub(pow(x1, p1), p2)))
+Simplified equation: (p1*p2**2 - x2)*(p2**2*x0**x1 - x1*(p0 - p2*x0))/(p1*p2**2*(p1 + x1 + x4))
+Raw equation: sub(div(mul(sub(mul(p0, p2), x0), x1), mul(pow(x0, x1), p2)), div(div(sub(x2, mul(p2, p1)), p2), add(add(x1, x4), p1)))
