+++ csvs/benchmarks/98_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-6.81268429 -6.85386658  0.44124254 -5.6475531  -3.97728797]
+Fitting parameters: [-5.93377153 -5.84750578  7.49699586  2.32626189  1.68142568]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (-x2 + 1)**x0/(x0*(p2*x2 - p2*(p1 + x1) - x3 + x4))
-Raw equation: div(pow(mul(div(sub(p1, p1), x2), x0), div(div(x3, x0), x3)), sub(add(sub(x4, mul(p2, x2)), x3), add(mul(x1, p1), p2)))
+Simplified equation: p2**2 - x2 - x3*(p0 + p1)*(x1 + x2*x4**2) + x3**x2
+Raw equation: sub(add(pow(sub(add(x3, p1), p1), x2), sub(mul(p2, p2), x2)), mul(add(mul(x2, mul(x4, x4)), x1), mul(add(p1, p0), x3)))
