+++ csvs/benchmarks/20_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-8.43689434  3.98046318  3.45850246 -6.95577587 -1.61425992]
+Fitting parameters: [ 5.33786049  8.26253482  4.31023205  4.52964535 -8.37286157]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: ((x0*x3 + x2)/x3)**p2*(p0 + x4*(p2*x1*x2 - x0 + x2))/(x4*(p0*p1 + p2))
-Raw equation: add(sub(mul(add(div(p0, x4), x2), x0), mul(mul(x1, p2), x2)), pow(add(div(x0, div(x2, x3)), p2), mul(add(p0, p1), p2)))
+Simplified equation: x2*((p0*(p0 + x4) - x2)/p0)**(-x3)*(-x3 + 1)/(x0*x1*x3)
+Raw equation: div(sub(mul(div(div(x2, x3), x2), p1), mul(x1, mul(p1, x0))), sub(pow(add(x4, p0), div(x2, p0)), div(mul(x3, p1), p1)))
