+++ csvs/benchmarks/67_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 9.73504485 -2.18586182 -4.83263919 -9.0720932   6.32724972]
+Fitting parameters: [ 1.77100809  9.94176218 -9.72074831 -3.62842293 -2.02811905]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: -x1**(-x0)*x2*x4**2*(p2**2 + x2)/(p2**2*(p0 - x4**2)*(p0 + p2 + x0))
-Raw equation: add(div(mul(div(div(x2, p2), p2), x2), add(add(p2, p0), x0)), mul(mul(sub(x4, div(p0, x4)), p2), div(pow(x1, x0), x4)))
+Simplified equation: -(-(x1*x3 - x2)/x1)**x3*(p1*x3 - x0)*(p0*(p1 - x0) - x4 + (p0 + x2)**x0)
+Raw equation: sub(div(mul(mul(pow(x2, x1), x3), x3), mul(sub(p1, x3), x0)), mul(sub(add(p0, sub(x0, p1)), x4), add(pow(p0, x2), x0)))
