+++ csvs/benchmarks/64_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-6.27869282  1.31484235 -5.9584723  -0.68919601 -7.54405375]
+Fitting parameters: [-0.43586524 -7.3623764  -5.11099869 -4.62342177 -4.44009235]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (x0*x2**2*(x1 + x2) + x0*x4 + x3*x4*(p1*x1*x3 - x3 + x4))/(x3*x4)
-Raw equation: add(sub(div(add(add(x0, x3), x4), x3), mul(mul(p1, x3), x1)), mul(mul(div(x2, div(x3, x0)), x2), add(div(x1, x2), x4)))
+Simplified equation: p1*x2**2 - x3 + (x2 + 1)**x0/(p0 + p1 - x3) + x4/(p1*x3)
+Raw equation: add(pow(div(add(div(x0, x0), x2), x0), sub(add(p1, x3), p0)), sub(div(add(x4, mul(p1, x3)), x3), mul(mul(x2, p1), x2)))
