+++ csvs/benchmarks/43_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-5.28342108  1.03334515  2.02083766 -6.628066    7.69539117]
+Fitting parameters: [-7.77598986  2.21384793 -6.09047308 -1.23687499  2.94727355]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: -(p2*(x1 - x3) + x0*x4*(x1 + x3))*((-(p1*x0 - p2)/x0)**p1)**(-(p0*x1 - p2)/x1)/x4
-Raw equation: sub(pow(mul(pow(div(p2, x0), p1), p1), sub(div(p2, x1), p0)), mul(sub(div(p2, sub(x3, x1)), x4), mul(add(x3, x1), x0)))
+Simplified equation: ((p2**2 - x1)/(p2*x2))**(x3**(p1 + 1))*(p0 + p2*x1**x1 + x4*(p0 + x0))
+Raw equation: mul(add(add(add(mul(x0, p0), x4), p0), mul(pow(x1, x1), p2)), sub(pow(div(p2, div(x1, p2)), x2), pow(mul(x3, p1), x3)))

[…]
