### 2019-08-15 f2d7ec7
redid the benchmarks preventing y_data be a constant
Former-commit-id: 06ff1833d40ea04c7143f8397212cbcce95d6757
