+++ csvs/benchmarks/81_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 4.67983056 -1.03296613 -3.27324345 -0.99844823 -3.650749  ]
+Fitting parameters: [-3.55072977 -8.55660656  0.87095958  1.78911682 -3.35565969]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (-p1*(-p0 + x2 + x4) + p2 + 1)*(p0 - p2 + x1 - x2 + x4 - 1)
-Raw equation: mul(add(mul(sub(sub(x2, x4), p0), p1), add(div(p2, p2), p2)), add(add(sub(x2, add(x4, x1)), p2), sub(div(x2, x2), p0)))
+Simplified equation: p1*(p2 - x4)**p0*(x1**p2)**x1 + x3*(-p0 + p1 + p2) + x3
+Raw equation: sub(sub(mul(mul(pow(p2, x4), p0), p1), pow(pow(x1, p2), x1)), mul(sub(sub(p0, add(p2, p1)), x3), add(sub(x4, x3), x4)))
