+++ csvs/benchmarks/5_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 0.37428786  1.35356458 -6.9573155   1.65112231  8.50423907]
+Fitting parameters: [-7.99772953  8.90233601 -0.36317751 -8.06437299  7.787516  ]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (p0*((p0 - x2 + x4)**p0)**((p2*x3 + x0)/p2) + p1*p2*(x1 + x4))/p0
-Raw equation: sub(add(pow(add(pow(x4, p0), x2), p0), div(add(x0, p2), x3)), div(div(add(x1, div(x0, x0)), x4), div(div(p0, p1), p2)))
+Simplified equation: ((p0*(-p1 + x2*(p1 + x0)**p0) + x4)/p0)**(2*x0 + 2*x2 + x4 - 1)
+Raw equation: add(mul(pow(pow(add(p1, x0), p0), x2), sub(div(x4, p0), p1)), sub(add(add(x0, div(x3, x3)), x4), add(add(x2, x2), x0)))
