+++ csvs/benchmarks/95_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 1.94896475  1.69368537  5.04038759 -9.43302682 -7.53685083]
+Fitting parameters: [ 0.1605796   6.74930166 -0.34151859 -7.00227709 -6.86680844]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (x0*(p2/x2)**p2/(p2*x1*x2))**(p0*(p1 + x0 - x2)*(x0 - x1 + x2))
-Raw equation: div(div(pow(mul(pow(p2, x2), p2), x0), mul(mul(x2, x1), p2)), add(mul(mul(x2, sub(x0, x1)), p0), sub(add(p1, x0), x2)))
+Simplified equation: p2*x2*x4**p2*(2*x2)**x4*(2*x1 - x3)*(x0*(-p1 + x0) + x3)/x0
+Raw equation: add(mul(sub(mul(div(x3, x0), x0), p1), mul(pow(x4, p2), x2)), mul(add(mul(x1, sub(x1, x3)), p2), pow(add(x2, x2), x4)))
