+++ csvs/benchmarks/42_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 4.52141753 -4.63051699  6.02542142 -1.13822371 -9.74845748]
+Fitting parameters: [ 3.50692796 -5.70771209 -2.14038669 -8.04774006 -1.7890844 ]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (-p1*p2*(2*x0 - x1) - x2*x4**2 + x3*x4*(x0 - x2*x3))/(x2*x4)
-Raw equation: sub(mul(sub(sub(div(x0, x2), x3), x3), add(sub(x4, p2), p2)), div(mul(mul(p1, div(p2, x2)), x4), sub(add(x0, x0), x1)))
+Simplified equation: -(p0 - p2*(p2 - x0))*(x0 - (p0 - x4)*(p0 - x3*x4**x0))*(x4 + 1)/(p2*(p0 - x4))
+Raw equation: mul(sub(add(div(mul(p0, p2), x0), p2), div(add(p1, p1), x4)), add(sub(div(x0, sub(x4, p0)), p0), pow(mul(x4, x0), x3)))

[…]
