### 2019-08-15 224e2b0
purge physical db files as well so we do not mix projects
Former-commit-id: f824098e9525f1fe955d3deb293626d51160b12c [formerly 36662c864d43046b7e9d39d00a7120bab7fa4fe1]
Former-commit-id: 6f93d42b2cb8e545fac70d8fe9237f6ee1be1ee9
Former-commit-id: 619a66f35033c2e370f7fa98c86d85e8578bb456
