### 2019-08-15 b9902cd
adding find_matching_SRGP_job command
Former-commit-id: b105641012c9e7c1a9eb020dc9922b7edc2ee264 [formerly 3a7c73d1541d61deaa716ed7577b98d5254bbb72]
Former-commit-id: cc01bae3445b925e26b11c6da0b177b6a6ab2e95
Former-commit-id: c98159ae274e2b281106757b2a52874a539669c6
