### 2019-08-15 6645dab
Update database.py
added note re dropbox_key

Former-commit-id: 075940e9318dadb7af04286fe08e0952f3e33601
