### 2019-08-15 ffbe765
Update database.py
select * -> select n_evals

Former-commit-id: c7b36d16bf6d1ad3da189cded432de0351447578
