### 2019-08-15 edcaaaf
Update database.py
Former-commit-id: 66b232e596bf8b3f3d850b07d98d1a4d360c0490 [formerly fe4c582fa889750b4ccb824eb2cd3a2a32b5fc75]
Former-commit-id: 5e7901d60a7564b0490e1a778e7bba404884c597
Former-commit-id: 3a8562e6ba615e435044cfe7c3740b4c791ffd06
