### 2019-08-15 5b3d022
Update and rename evaluate_SRURGS_on_benchmarks.py to submit_SRURGS_on_benchmarks_to_db.py
Former-commit-id: 2c65e49b6ca4e2f949e5f41204874ad52b92208f
