### 2019-08-16 a0904d1
Benchmarks giving error when calc np.std of y data
Former-commit-id: a956f6c69c947241b9ad0a7550c1bca532de89d5
