### 2019-08-16 217db6d
remade benchmarks, ensuring will load, and put badges locally
Former-commit-id: 476de82dfe1489b8f2b98789112ada7989c28773 [formerly b4793e4c61c4f140afe8fa28806ae03932ebbb11]
Former-commit-id: ce5b6c3c4618f744ae261e98ac00844795344f39
Former-commit-id: abfb889b4d4bbafd1af44c4aa4e16a7e4b54c949
