+++ csvs/benchmarks/54_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-7.07070172  8.88071281  9.17118606  3.75289083 -2.14701213]
+Fitting parameters: [ 0.73728468  9.56638792 -1.92981464 -0.62750063  2.42600125]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p1*p2*x1**(p1 - x4)*x2*(-p0 + p1)**(-x2)*(x2 + x3*(x0 - x4))/x3
-Raw equation: div(add(mul(mul(sub(x2, x3), x0), x4), pow(mul(x1, p1), p2)), mul(div(div(p1, pow(x1, x4)), x2), sub(pow(p1, p0), x2)))
+Simplified equation: ((p1 + x2*(p2 - x4))/x2)**(p1 + 1)/(p0 - p2 - x0*x4*(p0 - 1) + x4)
+Raw equation: sub(div(add(pow(div(p1, x2), p2), x4), div(add(p2, p2), p1)), sub(sub(mul(x0, mul(x0, p0)), x4), sub(sub(p2, x4), p0)))
