+++ csvs/benchmarks/90_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 9.04154309  7.19168834  7.33145972  7.30216958 -2.22672572]
+Fitting parameters: [-2.17597002 -1.83317665  3.96720704 -5.1375295   3.99716214]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p1**(-x3 - 1)*(p1**(x3 + 1)*x0*(x2 - x4)*(p0*(x3 + x4) + p0**x3 + 2*p1) - p2)/(x0*(x2 - x4))
-Raw equation: add(add(add(add(mul(x3, x4), p0), p1), add(pow(p0, x3), p1)), div(div(div(p2, pow(p1, x3)), p1), mul(sub(x4, x2), x0)))
+Simplified equation: -x4*(p2*(p1 + x1) - p2**(p2 - x1)*x1)*(p0 + p1 + x2)*(p0 - x2 + x4)
+Raw equation: mul(mul(sub(mul(add(x4, p0), x2), x4), add(add(p0, p1), x2)), sub(mul(pow(p2, sub(p2, x1)), x1), mul(add(p1, x1), p2)))
