+++ csvs/benchmarks/81_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-3.55072977 -8.55660656  0.87095958  1.78911682 -3.35565969]
+Fitting parameters: [-9.85681485 -0.01100049  3.35365132  2.85068555 -2.29099254]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p1*(p2 - x4)**p0*(x1**p2)**x1 + x3*(-p0 + p1 + p2) + x3
-Raw equation: sub(sub(mul(mul(pow(p2, x4), p0), p1), pow(pow(x1, p2), x1)), mul(sub(sub(p0, add(p2, p1)), x3), add(sub(x4, x3), x4)))
+Simplified equation: -(x3*(2*x2)**x2)**(x2*(p1 + x4))*(x1 + x2*(p2 - x4))/((p2 - x4)*(-p1 + p2 + x3))
+Raw equation: pow(add(mul(mul(pow(x2, x2), x2), x3), mul(add(x4, p1), x2)), div(sub(div(x1, sub(x4, p2)), x2), sub(add(p2, x3), p1)))
