### 2019-08-16 476de82
remade benchmarks, ensuring will load, and put badges locally
Former-commit-id: b4793e4c61c4f140afe8fa28806ae03932ebbb11
