+++ csvs/benchmarks/65_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-0.26675989 -7.7630108  -2.81029088  3.53317644 -4.26095533]
+Fitting parameters: [-6.30543345 -9.48159788 -4.17472171 -2.97517123 -8.13166486]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: x3*(-p1 + x0**(-p0 + x1))**(x1*(x1 + x2))*(x1 + 1)
-Raw equation: add(sub(mul(add(mul(x3, p1), p1), x1), div(mul(x3, p0), p0)), sub(pow(pow(x0, sub(x1, p0)), p1), mul(add(x1, x2), x1)))
+Simplified equation: (p1**2*x1 - x0**x0)*(-p2**2 + p2*x3*(-p2 + x2 + x4) + x3*x4)/(p1*x3)
+Raw equation: sub(div(mul(sub(add(p2, x3), p2), x4), div(add(x4, p2), x2)), mul(sub(div(p2, pow(x0, x0)), p1), mul(mul(x1, p1), p2)))
