+++ csvs/benchmarks/78_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-0.33278135  5.21047743  3.70413219  4.46618639  1.32530967]
+Fitting parameters: [ 0.38153399  0.08606304 -5.45063431 -2.00804371  4.62386677]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: -p2*x3*(p0*x4**2*(x0 - 1) + x0)/(p0*x0*x4*(p0*x1 + x3*(x0 + x2**p0)**p0))
-Raw equation: sub(sub(div(div(mul(x4, x0), x4), p2), div(div(p2, x4), p0)), add(add(pow(x0, pow(x2, p0)), p0), mul(div(x1, p0), x3)))
+Simplified equation: -p0**(p1 + 1) + x3*(-p0 + x4)**x3 - x4*(x4/(x2*x3))**x4
+Raw equation: sub(sub(mul(sub(pow(x4, p0), x3), x3), mul(pow(p0, p1), p0)), pow(div(mul(x4, mul(x2, x3)), x4), div(mul(p0, p0), x4)))
