+++ csvs/benchmarks/75_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 2.44054898  0.67242677 -4.03575495 -3.38876108  3.71588949]
+Fitting parameters: [ 9.97023531  8.90488951 -0.18664786  9.93697842  9.11946413]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: -p1*x0**(-p1)*x4/x1 - p2*x1**(p0**p2)*(x4**p0)**(-x4) + x0**(-p1)*x4/x1
-Raw equation: sub(sub(div(mul(div(x2, x2), p1), x4), pow(mul(x0, p1), x1)), mul(div(pow(x1, pow(p0, p2)), p2), pow(pow(x4, p0), x4)))
+Simplified equation: -2*p0*(p0 - x3*x4)/(x3*(-p1*(p2 - x1*x3) + p1 + x3 + x4))
+Raw equation: mul(sub(div(mul(div(p0, x4), x3), p0), div(add(x4, x4), x3)), sub(sub(mul(p2, mul(x3, x1)), p1), add(add(x4, p1), x3)))
