+++ csvs/benchmarks/52_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 4.65382371  1.40096293  4.74450466  8.33383849 -2.95051734]
+Fitting parameters: [-1.64405867 -4.2922465   0.4316745   9.43324319  9.07710904]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p0 + p0/p1 - p1 + x0 + x2*x4**x2*(p1 - 2*x1)**x3 + 2*x3
-Raw equation: add(add(add(add(div(p0, p1), x0), x3), add(sub(p0, p1), x3)), sub(mul(pow(p1, add(x1, x1)), x3), mul(pow(x4, x2), x2)))
+Simplified equation: p2**p2*x3 - x0*(p0 - p1 + x1) + (x0 + x1)**p1*(p1*x0 + x4)
+Raw equation: add(mul(mul(add(pow(x0, x1), p1), x0), div(add(x4, x0), p1)), add(sub(mul(p1, add(x1, p0)), x0), pow(mul(p2, p2), x3)))
