### 2019-08-16 af0bd3b
remade benchmarks, ensuring will load, and put badges locally
Former-commit-id: 80359610a1f73d43de592d25740a9eadf85436d3
