+++ csvs/benchmarks/74_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 4.51902124 -4.10613135 -1.44635028 -0.18155591  3.20562483]
+Fitting parameters: [-7.63026641  1.95179662  5.08897853  6.68080519  8.74900328]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (p1*(p2 - x0)*(x1**2 + x2) - x1**2*x2**(x1*x2 + 1)*(p0*x2 + x4))/(p1*x1*x2*(p2 - x0)*(p0*x2 + x4))
-Raw equation: div(div(add(div(add(x2, x1), x1), x2), add(mul(x2, p0), x4)), pow(div(mul(x2, mul(x2, x1)), x1), sub(mul(x0, p2), p1)))
+Simplified equation: -p2*x0*x4*(p0 - x3*(x1 - x2))*(p1*p2 - x4)/(p1*x3*(x1 + x4))
+Raw equation: mul(div(sub(add(mul(p0, x3), x2), x1), sub(div(x4, p1), p2)), div(mul(mul(p2, add(x4, x1)), x0), sub(add(x4, x2), x2)))
