+++ csvs/benchmarks/73_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-8.683701   -2.88148167  3.18275301 -8.86205583  2.63590202]
+Fitting parameters: [ 0.48576107 -8.69343283  4.78981315  3.75703832  5.41745233]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (p2 - p2**x0)/(x2*(p1 - x4**x3)) - (p0 - p1/x0)/(p1*(p2*x2 - x4))
-Raw equation: div(div(sub(div(sub(p1, x0), p0), p1), sub(mul(x2, p2), x4)), sub(div(div(p2, pow(p2, x0)), x2), sub(pow(x4, x3), p1)))
+Simplified equation: p1*(x4**x3)**(-p1)/x1 - x0*(x4**x3)**(-p1) + x1*(x4**x3)**(-p1) - (-p1 + x0/(p2*x4))**(p2*x2*x3)
+Raw equation: div(div(add(sub(sub(p1, x1), x1), x0), pow(pow(x4, x3), p1)), sub(pow(div(x0, mul(x4, p2)), p1), mul(mul(x3, x2), p2)))
