+++ csvs/benchmarks/8_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 2.98243223 -4.23159031 -8.5398977   9.3009809   0.01245517]
+Fitting parameters: [-8.90033644 -0.50632634 -7.20003476 -1.96419864  3.12545073]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: (((x1 + x3*(p1 - x4))/x3)**x3)**((-x3*(p2 + x0*x2) + 1)/x3)
-Raw equation: sub(add(pow(pow(div(x1, x3), p1), x4), add(sub(x3, p1), p1)), pow(div(sub(x3, sub(x0, x0)), x3), add(mul(x0, x2), p2)))
+Simplified equation: p0*x1*(x2 + x4)*(p0*x0**2 + ((p2*x1 + x0)/p2)**x1)*(p1*x4 - x3)/p1
+Raw equation: div(add(add(mul(pow(x0, p2), x1), x1), mul(mul(x0, p0), x0)), mul(mul(sub(x4, div(x3, p1)), x1), add(mul(x4, x2), p0)))
