+++ csvs/benchmarks/55_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-9.47395757  4.66047612  6.84613899 -3.82262833  6.27789474]
+Fitting parameters: [5.6434889  5.47006202 4.99341183 0.9146791  8.37973369]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: x1**2/(x4*(p2 - x4)*(x0 - x4)) + (p1/x0 + x3)/(p1 + p2)
-Raw equation: add(div(mul(add(div(p1, x0), x3), x2), mul(add(p2, p1), x2)), mul(div(div(x1, sub(x4, p2)), x1), sub(mul(x4, x0), x4)))
+Simplified equation: p0**2*(x1/p0)**x2*(p0**p0 - x2)/(x3*(p1 + x2)*(x2 + x3 - x4))
+Raw equation: div(div(mul(mul(pow(x1, p0), x2), p0), mul(add(p1, x2), x3)), div(sub(mul(x2, pow(p0, p0)), p0), sub(sub(x4, x3), x2)))
