+++ csvs/benchmarks/98_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-5.93377153 -5.84750578  7.49699586  2.32626189  1.68142568]
+Fitting parameters: [-5.99956857 -8.02644745  8.12224275  6.56074122 -6.46159284]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p2**2 - x2 - x3*(p0 + p1)*(x1 + x2*x4**2) + x3**x2
-Raw equation: sub(add(pow(sub(add(x3, p1), p1), x2), sub(mul(p2, p2), x2)), mul(add(mul(x2, mul(x4, x4)), x1), mul(add(p1, p0), x3)))
+Simplified equation: -p2**(-p1)*(x2*(x1 - x2) + x4*(-p2 + x4)**x0)*(p0*p2**p1*x2 - p0*x1*x3 + p2**(p1 + 1))/p0
+Raw equation: mul(sub(sub(pow(mul(x4, p2), x0), x4), mul(sub(x2, x1), x2)), div(mul(sub(x1, pow(p2, p1)), x3), add(div(p2, p0), x2)))

[…]
