### 2019-08-16 afd27bf
remade benchmarks, ensuring will load, and put badges locally
Former-commit-id: af0bd3b3f19a1dc5fd89031a75684737141fe12a [formerly 80359610a1f73d43de592d25740a9eadf85436d3]
Former-commit-id: c35d6756a77a6f9b4511d1c0e51ef000b9d980ff
Former-commit-id: f7cb7a4a0787092bd18362a9e0853bc277dd7553
