+++ csvs/benchmarks/77_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 9.91617348 -7.57160212  8.41653289  5.53424671  7.61674358]
+Fitting parameters: [-5.97734976 -8.0098862  -6.49512844  1.2116561   5.67955749]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: -(p2 + 2*x3)**x3*(x2 + x3)**p1 + (-x1 + x4/x3)**x2*(x1 + x3)**p0
-Raw equation: mul(div(sub(sub(pow(x4, x3), x1), x2), add(pow(x3, x1), p0)), add(pow(mul(x3, add(p2, x3)), x3), pow(add(x3, x2), p1)))
+Simplified equation: (p0*x1*x4 - p1**2)*(x1 + x2**x0 + (x0*(x2 + 1))**x4)/x4
+Raw equation: mul(add(sub(sub(mul(x3, x1), x3), p0), mul(div(p1, x4), p1)), add(pow(add(x0, mul(x0, x2)), x4), pow(add(x2, x0), x1)))
