+++ csvs/benchmarks/93_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-4.02091017  0.16762101  8.76017671  9.66716245  1.11690997]
+Fitting parameters: [ 7.02611926 -1.83407774 -3.41370579  5.5671781   0.97196588]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: x4*(-p0 + x1)**p1*(p0 - p2)*(x4 + 1)*(x0 + x4 + 1)/p0
-Raw equation: mul(mul(mul(add(div(p1, p1), x4), x4), pow(sub(x1, p0), p1)), add(add(mul(x4, div(x0, x0)), x0), sub(div(p0, p2), p0)))
+Simplified equation: p2*x4*(x0 + x1*(p0 - x0 + x3))*(-p2 + x2 + x4)**x3/(p0*x1)
+Raw equation: mul(add(sub(pow(mul(x2, x4), p2), x3), div(mul(p2, p0), x4)), sub(add(add(p2, add(x0, p2)), p0), div(add(x0, x1), x3)))
