+++ csvs/benchmarks/24_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-7.11401471 -8.29715971  2.38266952 -2.36659521 -8.70920989]
+Fitting parameters: [ 1.28475783 -4.55755885  4.73632829 -2.35399031 -6.30263394]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: p2**(x0 + 1)*(p2/(x1*x4))**x0*(-x2*(p1 + x0) + x4*(-p1 + x1)**x0)
-Raw equation: sub(mul(sub(mul(pow(x1, p1), x0), x4), add(mul(x0, p1), x2)), mul(div(pow(p2, mul(x1, x4)), x0), pow(mul(p2, x0), p2)))
+Simplified equation: ((x4*(-x0 + x1)**p0)**((x3 + x4)**x1))**((-p0*(x0 - x2) + x2)**(p0*x0*x1))
+Raw equation: sub(mul(pow(pow(pow(x1, x0), p0), x4), pow(add(x4, x3), x1)), add(mul(pow(p0, sub(x2, x0)), x2), mul(mul(p0, x0), x1)))
