+++ csvs/benchmarks/27_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-4.98213097  4.487349   -8.63919994 -0.68122136 -0.05568714]
+Fitting parameters: [-5.82420488  8.34021607 -2.04227891  7.74867687 -4.75674247]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: x0*(x0 + 1)*(p2*x1 + p2*x4*(p0 - x4) - x2**x3*x4)/(p2*x4)
-Raw equation: add(sub(sub(div(mul(x1, x4), p0), x4), pow(div(x2, x3), p2)), mul(add(div(x0, mul(x0, x0)), x0), div(mul(x2, x2), x0)))
+Simplified equation: ((-p0*x0 + x2**2*(x0 - x1)**p0)/x2)**(-0**x1 + p0*x3 + p1 - x2)
+Raw equation: sub(pow(pow(sub(mul(x0, x1), p0), x2), mul(div(p0, x2), x0)), add(sub(sub(p1, mul(x3, p0)), x2), pow(sub(p0, p0), x1)))
