+++ csvs/benchmarks/50_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [-4.84032328 -4.09679627 -4.58611061 -3.63805931 -6.17686759]
+Fitting parameters: [ 1.62628329 -8.15879078 -0.30301909 -5.29273887  2.92619294]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: x3**2*(p2 - x1)*(p1*x0 - p2)*(1/x3)**p0/(p1*p2**2*x1)
-Raw equation: div(mul(mul(div(sub(p2, p1), x0), x3), mul(mul(x1, p2), p2)), mul(pow(div(x4, mul(x3, x4)), p0), sub(mul(x1, p2), x3)))
+Simplified equation: -p1 + x0 - x1*(p0 - x2)**x2*(p2 - x2 + x4) + x3 - (x3**(p1 + p2))**p1
+Raw equation: mul(mul(sub(sub(pow(p0, x2), x2), x1), sub(sub(x2, p2), x4)), pow(sub(pow(x3, add(p1, p2)), p1), add(sub(x0, x3), p1)))
