+++ csvs/benchmarks/79_params.txt
@@ -1,6 +1,6 @@
 Permitted variables: ['x0', 'x1', 'x2', 'x3', 'x4']
 Permitted fitting parameters: ['p0', 'p1', 'p2', 'p3', 'p4']
-Fitting parameters: [ 3.78498973  1.10812432 -2.19818239  0.69544236 -4.93349168]
+Fitting parameters: [ 1.51113297 -8.04757313  2.8576419  -0.52498505  2.05495146]
 Permitted functions: ['add', 'sub', 'mul', 'div', 'pow']
-Simplified equation: -p0**(-x0)*x0*(p0**x0*x2 - x0)*(p0*x1*(x1 - x4) - x2*(x0 + x1*x2))/x1
-Raw equation: add(mul(mul(add(div(x0, x1), x2), x2), mul(sub(x4, x1), p0)), sub(div(mul(x0, pow(p0, x0)), x2), sub(sub(x2, x2), x0)))
+Simplified equation: -p0*x1*x3**p2*(p0 + x4)*(p0 - p1*x0)*(x2*x4 - x3)/p2
+Raw equation: sub(mul(mul(mul(div(x3, x4), x2), p0), mul(pow(x3, p2), x1)), sub(mul(mul(p0, mul(p1, x0)), x4), div(add(p0, x4), p2)))

[…]
