### 2019-08-16 676187f
made the badges except travis-ci locally hosted
Former-commit-id: afee28373def180f47a00a48cc2f704e264e40a1
