### 2019-08-16 16c7df2
Update README.md
removed performance claim as not done comparison testing yet

Former-commit-id: 77bad99c767ec577923bc53d252d0b078be0426d [formerly a1bad56c2d81e03b85063e492d5059da1e403e4f]
Former-commit-id: 6e4da66a76f93046b851313fcf5f8fee681a0ded
Former-commit-id: ec79c42ec1139d88ddd78801b565938cd853b703
