+++ pySRURGS.py
@@ -27,6 +27,11 @@ from sqlitedict import SqliteDict
 import collections
 from itertools import repeat
 import multiprocessing as mp
+import matplotlib 
+matplotlib.use("TkAgg")
+import matplotlib.style
+matplotlib.style.use('ggplot')
+import matplotlib.pyplot as plt
 
 class SymbolicRegressionConfig(object):
     ''' 
@@ -747,12 +752,17 @@ class Enumerator2(object):
             j_i = left_j_i + right_j_i
         return j_i
 
-def create_fitting_parameters(max_params):
+def create_fitting_parameters(max_params, param_values=None):
+    # param_values must be None or a np.array of length max_params
     params = lmfit.Parameters()
     for int_param in range(0, max_params):
         param_name = 'p'+str(int_param)
         param_init_value=np.float(1)
         params.add(param_name, param_init_value)
+    if param_values is not None:
+        for int_param in range(0, max_params):
+            param_name = 'p'+str(int_param)
+            params[param_name].value = param_values[int_param]
     return params
 
 def eval_equation(params, function_string, my_data, mode='residual'):
@@ -767,6 +777,10 @@ def eval_equation(params, function_string, my_data, mode='residual'):
         residual = eval(eval_string)
         output = residual
     elif mode == 'y_calc':
+        y_value = eval(function_string)
+        output = y_value
+    elif type(mode) == dict:
+        df = mode
         y_value = eval(function_string)
         output = y_value
     # if model only has parameters and no data variables, we can have a
@@ -1080,9 +1094,7 @@ def create_db(path_to_csv, additional_name=None):
         db_name = './db/' + csv_name + '.sqlite'   
     return db_name
 
-def compile_results(path_to_db, path_to_csv, SRconfig):
-    # reads the generated .sqlite file to determine the best models, then 
-    # prints them to screen!
+def get_resultlist(path_to_db, path_to_csv, SRconfig):
     (_, _, _, _, _, dataset, _, _, _) = setup(path_to_csv, SRconfig)
     result_list = ResultList()
     with SqliteDict(path_to_db, autocommit=True) as results_dict:
@@ -1090,9 +1102,45 @@ def compile_results(path_to_db, path_to_csv, SRconfig):
         for eqn in keys:
             result = results_dict[eqn]
             result_list._results.append(result)
-    result_list.sort()    
+    return result_list, dataset
+    
+def compile_results(path_to_db, path_to_csv, SRconfig):
+    # reads the generated .sqlite file to determine the best models, then 
+    # prints them to screen!
+    result_list, dataset = get_resultlist(path_to_db, path_to_csv, SRconfig)
+    result_list.sort()
     result_list.print(dataset._y_data)
     return result_list
+
+def plot_results(path_to_db, path_to_csv, SRconfig):
+    # reads the generated .sqlite file to determine the best model, 
+    # then plots it against the raw data. saves the figure to plot.png    
+    result_list, dataset = get_resultlist(path_to_db, path_to_csv, SRconfig)
+    if len(dataset._x_labels) != 1:
+        raise Exception("We only plot univariate data")
+    result_list.sort()
+    best_model = result_list._results[0]
+    param_values = best_model._params
+    params_obj = create_fitting_parameters(SRconfig._max_num_fit_params, 
+                                           param_values=param_values)
+    equation_string = best_model._equation    
+    evaluatable_equation_string = equation_string
+    eval_eqn_string = clean_funcstring(equation_string)        
+    data_dict = dict()
+    xlabel = dataset._x_labels[0]
[…]
