### 2019-08-16 21f7d3d
Update README.md
showing example performance plot

Former-commit-id: 377c5b9084fc9b3cd47edf6c1334ff1e11fe25a0
