+++ README.md
@@ -124,6 +124,8 @@ Running in multi processor mode
 
 ```
 
+[![Example performance](image/plot.svg)]()
+
 ## Author
 
 **Sohrab Towfighi**
