### 2019-08-16 bb19d67
Update README.md
added to man docs the -plot output

Former-commit-id: 6ee86e8606620114f88513a94e68da3921ec37c9
