+++ README.md
@@ -54,11 +54,11 @@ python3 pySRURGS.py -h
 The above command should render the following:
 
 ```
-usage: pySRURGS.py [-h] [-run_ID RUN_ID] [-single] [-count]
-                   [-benchmarks] [-funcs_arity_two FUNCS_ARITY_TWO]
+usage: pySRURGS.py [-h] [-run_ID RUN_ID] [-single] [-count] [-benchmarks]
+                   [-funcs_arity_two FUNCS_ARITY_TWO]
                    [-funcs_arity_one FUNCS_ARITY_ONE]
                    [-max_num_fit_params MAX_NUM_FIT_PARAMS]
-                   [-max_size_trees MAX_SIZE_TREES]
+                   [-max_permitted_trees MAX_PERMITTED_TREES] [-plotting]
                    train iters
 
 positional arguments:
@@ -96,6 +96,9 @@ optional arguments:
                         of the equation, increasing this number tends to
                         increase the complexity of generated equations
                         (default: 1000)
+  -plotting             plot the best model against the data to
+                        ./image/plot.png and ./image/plot.svg - note only
+                        works for univariate datasets (default: False)
 ```
 
 ### Important details
