### 2019-08-16 6d8117b
changed help string for -plot option
Former-commit-id: 931635b59ed604e92abc607f7c14821318c187c6 [formerly a5dfa987ecb19bfd9d2c0f9cadd618402d93fc5d]
Former-commit-id: 7c6fa8e9ad33d84aeecf243782bdd8c6ef5ee131
Former-commit-id: 3be39801c934b47900d6cac219afb910f87d6dce
