### 2019-08-16 ee9b105
reorganized -plotting option
Former-commit-id: be5a431612ef83aff29bd9156023450188849728 [formerly 47702b3078af156fec601687428c99201cb5d777]
Former-commit-id: c88e44c7f50e92e1faca841359670fb8709e3008
Former-commit-id: 3208241fe8eff8d9cf2e326266b56f2b5fa0408d
