+++ README.md
@@ -127,7 +127,7 @@ Running in multi processor mode
 
 ```
 
-[![Example performance](image/plot.svg)]()
+[![Example performance](image/plot.svg)]
 
 ## Author
 
