### 2019-08-16 fd5f67f
delete contents of db folder before running
Former-commit-id: b3538d7b903103f926d11d8dca3f0d071526802c [formerly 9ba81aeac05404243801f277768d61605fe1d670]
Former-commit-id: a9263d34ace4872a9bcef1ffc02dc9f88a7f87ed
Former-commit-id: 53ba4530094ab7c6ce255ae337cde9848de3e369
