### 2019-08-17 ed1148c
latest version for submitting SRURGS jobs
Former-commit-id: 9e6d5a2d004622c9cb6d14cccfac8a15e66485cd
