+++ experiments/submit_SRGP_on_benchmark_jobs_to_db.py
@@ -41,7 +41,7 @@ def generate_list_of_experiments(SR_config, start_index, count_experiments, n_ru
     popsize = 500
     numgens = 40
     # first twenty problems
-    for z in range(start_index,count_experiments):
+    for z in range(start_index,start_index+count_experiments):
         train = '$PYSRURGSDIR/csvs/benchmarks/'+str(z)+'_train.csv'
         for j in range(0,n_runs):
             run_ID = str(j)
