### 2019-08-17 ae80b2c
need to rewrite database to allow for SRURGS jobs then write continuously writing program
Former-commit-id: 7644dcf3ee753a3ae446c54a1c8a174f158fb1f9 [formerly d4dfaacc713e87118f281c1d6154f3b4ec37af15]
Former-commit-id: e67f5c0486c14548bd5da224dace4277c66ffa8c
Former-commit-id: aa9b3e1f37a7680fec6b0fa1e1d66f57f941b11e
