+++ README.md
@@ -42,6 +42,7 @@ Clone the repo then install the prerequisites.
 
 ```
 git clone https://github.com/pySRURGS/pySRURGS.git
+cd pySRURGS
 pip install -r requirements.txt
 ```
 
