### 2019-08-17 e3df124
experiments readme
Former-commit-id: 1f687bea2da73ccdba3960f9a3c1a298c7c0bd4d [formerly c2384418edc3637311f8935720456194014bf3a6]
Former-commit-id: 3f399a9ca0e80f086c215014c7ca4f8a7a4faac2
Former-commit-id: ebfee5c3ab64d8a2c4311caf65d03d0446dbff3c
