+++ experiments/readme.txt
@@ -1,0 +1,5 @@
+first we submit the SRGP jobs
+run them
+then we submit the SRURGS jobs with matching n_evals
+run them 
+then we submit jobs for FFX and ITSR
