### 2019-08-17 4182d37
revised code so that job_ID key is now the db name. should be faster to submit srurgs jobs
Former-commit-id: 1318e414efae5da0ca1d927bf67192f8935d1293 [formerly 204fe0217fb99c30531185fefe8b674f9d4f72ae]
Former-commit-id: 4710ac22f46e3398366298e2675e7128b1f3f965
Former-commit-id: ec5592cd6184c7e0f1e7d531343305cf2aa428f8
