+++ experiments/requirements.txt
@@ -1,3 +1,4 @@
 dropbox
 sshtunnel
 pymysql
+scoop
