### 2019-08-18 8804339
Update README.md
make the install script sudo requirement proof

Former-commit-id: 2466d13684b96c8da26baff5a42c7b3058bce5c4 [formerly ddd3cf4964855454d4f8c50bd86ddcef2c92054f]
Former-commit-id: 3d2192352fd558da195c006d565c26732c80255d
Former-commit-id: e61ac71d2e88aedc01255f6b168faff75208115d
