+++ experiments/database.py
@@ -51,6 +51,20 @@ def get_db_path(arguments_string):
             return db_name
     raise Exception("Not able to find db")
 
+def run_select_qry(sql):
+    with sshtunnel.SSHTunnelForwarder(('ssh.pythonanywhere.com'),
+        ssh_username=PYTHONANYWHERE_USERNAME, ssh_password=PYTHONANYWHERE_PASSWORD,
+        remote_bind_address=(DATABASE_HOSTNAME, 3306)) as tunnel:
+        mydb = pymysql.connect(user=PYTHONANYWHERE_USERNAME, 
+                                                 password=DATABASE_PASSWORD,
+                                                 host='127.0.0.1', 
+                                                 port=tunnel.local_bind_port,
+                                                 database=DATABASE_NAME)
+        mycursor = mydb.cursor()
+        mycursor.execute(sql)
+        myresult = mycursor.fetchall()
+        return myresult
+
 def submit_job_to_db(algo_argu_list):
     # job_ID is the database file name 
     num_job = len(algo_argu_list)
@@ -71,14 +85,14 @@ def submit_job_to_db(algo_argu_list):
             create_db_command = '''CREATE TABLE IF NOT EXISTS jobs(
                job_ID VARCHAR(200) NOT NULL,
                algorithm VARCHAR(200) NOT NULL,
-               arguments VARCHAR(200) NOT NULL,
+               arguments VARCHAR(400) NOT NULL,
                finished TINYINT(1) NOT NULL,
                n_evals INT NOT NULL,
                PRIMARY KEY ( job_ID )
             );'''
             mycursor = mydb.cursor()
             mycursor.execute(create_db_command)
-            sql = "INSERT INTO jobs (job_ID, algorithm, arguments, finished, n_evals) VALUES (%s, %s, %s, %s, %s)"
+            sql = "INSERT IGNORE INTO jobs (job_ID, algorithm, arguments, finished, n_evals) VALUES (%s, %s, %s, %s, %s)"
             db_path = get_db_path(arguments)
             val = (db_path, algorithm, arguments, 0, -1)
             mycursor.execute(sql, val)        
@@ -210,8 +224,8 @@ def run_all_SRURGS_jobs(placeholder):
             output_db = job_arguments[2]
             if (job_arguments[0] != pySRURGS_dir+'/pySRURGS.py'):
                 raise Exception("SQL injection?")
-            try:
-                job_arguments.append(str(n_evals))
+            try:                
+                pdb.set_trace()
                 sh.python(*job_arguments, _err="error.txt")
             except:
                 print(sh.cat('error.txt'))
