### 2019-08-18 e83b905
typo in run_SRGP
Former-commit-id: 5127a7c727a2ea4677c9d33560064556b5c031bf
