### 2019-08-18 13d1868
fixed issue with output db
Former-commit-id: a275bbe93ee3c5b9c4b231eee904e0663ba88c28 [formerly a515db7042ac7bb46fefb94170ad47968526ff4a]
Former-commit-id: 8841b813e54acaeef59371ca1cf58029ce462bcf
Former-commit-id: 8575b4987f421863bf5b883f060dd3ed178a2ed1
