### 2019-08-18 ed52616
ssigning n_evals for the SRURGS cases.
Former-commit-id: abf1a36ed6bd622f9712a9403c9dc4dcd287caa5 [formerly caa68923066048aafb7a824051aa3b38eb07101b]
Former-commit-id: 2bb80d8ecfa38714053f4c5f7ca560ee4b623606
Former-commit-id: fc29a20d477613e55012cbe2e41e6dca4603fb6c
