### 2019-08-18 98e3f9e
ssigning n_evals for the SRURGS cases.
Former-commit-id: 11b95b3b868210d5d241328bf2d78c3ab28f9f41 [formerly a3df3070aab138ba8d2acde694396f8977e8506e]
Former-commit-id: c488316840c60abc0ee4686b684b6ad1c88a11c5
Former-commit-id: 32adf0bc6414a488ebfb3430131998477aef65d2
