### 2019-08-18 9497de5
removed part where i set n_evals to 20 for testing
Former-commit-id: 234a64d5d9e14657f63ead65e6d327b5b9c9bd1d [formerly 50b3e09d7859ecee48ebf7013bcb9428dc162cd7]
Former-commit-id: 8976c69e5969a4cc22900c2788a93f1c8da96764
Former-commit-id: a3d43309f33f39d51b391144aa1d58f0ff05a31c
