### 2019-08-18 2efc497
fixed issue with finished = 1 cases. will test
Former-commit-id: 7c51ed8ec22e104ea871a1bf43fa54151c07de1f
