### 2019-08-18 f6fb99d
making memoize an option, so as to not mandate for low ram computers
Former-commit-id: 5d3d72ee1ab95ca4351ca293f144299b0feb112d
