### 2019-08-18 662b660
making memoize an option, so as to not mandate for low ram computers
Former-commit-id: f6fb99d0b5be25d0ba795679c88b47196883f259 [formerly 5d3d72ee1ab95ca4351ca293f144299b0feb112d]
Former-commit-id: f7618516dfa19721195eb8d153b168594d2d0c94
Former-commit-id: f002e5d52511dc340fcc2024d160bcab7374cc83
