### 2019-08-18 1c26456
fixed minor issue in plot command
Former-commit-id: 49dc24d9ed5eab4ea43f46ac4e96d1716b28d4d4
