### 2019-08-23 #1675 @sohrabtowfighi
Hi, I have rewritten the test script `test.py` to better meet criteria described herein.
