### 2019-08-26 pySRURGS/pySRURGS#1 @henrik227
Unexpected results for simple test case (JOSS review)
For the JOSS review (https://github.com/openjournals/joss-reviews/issues/1675), I created a very basic test case to try pySRURGS. This is my data:

```
x, y
1, 2
2, 3
3, 4
4, 5
5, 6
6, 7
```

With `python pySRURGS.py -max_num_fit_params 1 -max_permitted_trees 2000 -funcs_arity_two=add,sub,mul,div -plotting ./csvs/test.csv 100`, I get the following output:

```
Running in multi processor mode
105it [00:14,  7.06it/s]
Making sure we meet the iters value
  Normalized Mean Squared Error       R^2  Equation, simplified                                           Parameters
-------------------------------  --------  -----------------------------------------------------------  ------------
                    0            1         p0*(p0 - x**2)                                                   1
                    6.68694e-17  1         -(p0 + x**2*(x - 1))*(p0**2 - x**2)/(p0*x**4)                    1.03e-08
                    0.00811055   0.999165  (-p0**3*(x + 1) + x*(2*p0 + 1)*(p0**2 - 1))/(x*(p0**2 - 1))      0.57
                    0.0209102    0.997802  (p0 + x)*(p0 - x + 1)/(x**2 + (p0 + x)*(2*p0 + x))               1.54
                    0.042286     0.996303  p0*(x*(x - 3) + 1)/x                                            -0.381
```

Obviously, `1 - x**2` is not a good model for this data, yet still it shows up as the best solution. However, the plot looks fine:
![plot](https://user-images.githubusercontent.com/25176271/63683854-c9ed4680-c7fb-11e9-90bc-4cd5d9b8a687.png)
Am I doing something wrong, or is this a bug?
