### 2019-08-26 pySRURGS/pySRURGS#1 @sohrabtowfighi
In the $pySRURGS_dir/db directory, there should be a file called test.db, corresponding to the name of your csv file. Can you share that with me?
