### 2019-08-29 pySRURGS/pySRURGS#3 @henrik227
Documentation of command line arguments
I think the documentation of the command line arguments could be improved, specifically:
- What is the purpose of the `-run_ID` argument? What would the ID be used for?
- What is the use of counting all possible equations (with `-count`)? For example, does this number help me to decide what to use for `-max_permitted_trees`? Also, isn't it possible to have an infinite number of possible equations?
- The help says that the default for `-path_to_db` is `None`, but even if it's not set, something is stored in the `pySRURGS/db/`.
- In general, what exactly is stored in the database file, and in which format? You could add suggestions for how to read this data.

If this ends up making the command line help too long, you could also provide a more detailed description in the `README.md` (instead of copying the command line help), and use shorter descriptions for the command line.

This issue is part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/1675).
