### 2019-08-29 pySRURGS/pySRURGS#3 @pySRURGS
> What is the purpose of the -run_ID argument? What would the ID be used for?

I removed -run_ID as it was extraneous now that we have -path_to_db.

> What is the use of counting all possible equations (with -count)? For example, does this number help me to decide what to use for -max_permitted_trees? Also, isn't it possible to have an infinite number of possible equations?

Indeed, -count can give you a ballpark idea regarding the effect of changes in the problem configuration (eg: going from sin,exp,tan -> sin,exp,tan,cos,sinh,cosh) on the solution space. It would be possible to have an infinite number of equations, but we limit the search space using the -max_permitted_trees parameter, making finite the number of equations.

> The help says that the default for -path_to_db is None, but even if it's not set, something is stored in the pySRURGS/db/.

I added to the doc string to explain that if path_to_db is not set, then the path_to_db is set based on the csv filename.

> In general, what exactly is stored in the database file, and in which format? You could add suggestions for how to read this data.

I have added a separate section in the readme.md explaining how to access the entries of the database file.
