### 2019-09-01 #1675 @sohrabtowfighi
I have added an -exhaustive flag which performs brute-force exhaustive search instead of uniform random global search
