### 2019-09-05 pySRURGS/pySRURGS#6 @SirSharpest
JOSS review comments 
# Review performed using most-current github (not in releases)
- https://github.com/pySRURGS/pySRURGS.git

# Installation
- Installation instructions work well
  - Though, it maybe worth while putting on pip as a package?

# Testing
- the function `test_command_line_code` shows that tests are done to command-line interface, however this doesn't give an indication to the results of the tests.
  - In other words: a test could be run, instantly return 0 and be considered a "success".
- I wonder, would it be possible to provide a testing table with expected outputs. This way a user can verify that their usage is running as expected

# Documentation
- The URL for documentation https://pysrurgs.github.io/pySRURGS/ doesn't come across as very intuitive, perhaps usage of a documentation generator e.g. Sphinx would do well.
- More clarity could be given in the explanation of results (see below confusion of mine).

# Methods
- I haven't dug into the methods too heavily but is it possible to ensure somehow that for given input you always get the same output, the stochasticity of the outputs? I noticed this when my results of your readme.md example were different.

# Results
- How should one interpret the output
    - is `(-p0*p1 + p2 + x**2/p0)**(-x*(p1**p1)**x)` equivalent to ![this](http://www.sciweavers.org/upload/Tex2Img_1567690636/eqn.png)
      - The pX is the coefficients given as parameters?
- If the above point is understood correct, then it results are not given in their most simplified form (though, I admit this may be a technical misunderstanding of mine)

# Other
- As changes were made recently, the releases of the software is now out of date (at time of posting this comment)
- Perhaps I've overlooked it, but was this work funded or supported by any research councils that need mentioned?
- The acknowledgements cited in the readme.md should be done better
     - You should indicate the artists for the icon, as well as the original location that you took it from, as well as any license which is applied to it
     - The algorithm of Tychonievich, 2013, seems to be a novel addition to your software, did they have enough input in this project to warrent authorship also?
