### 2019-09-05 #1675 @SirSharpest
I have added my inital comments [here](https://github.com/pySRURGS/pySRURGS/issues/6), though I am happy to elaborate if I have been unclear.
