### 2019-09-05 pySRURGS/pySRURGS#6 @SirSharpest
This is referencing [JOSS submission](https://github.com/openjournals/joss-reviews/issues/1675)
