### 2019-09-05 pySRURGS/pySRURGS#6 @sohrabtowfighi
> would it be possible to provide a testing table with expected outputs

I will work on the command line tests using this idea

> The URL for documentation https://pysrurgs.github.io/pySRURGS/ doesn't come across as very intuitive, perhaps usage of a documentation generator e.g. Sphinx would do well.

I am using pydoc, a documentation generator that is part of the python standard library. It may not be as aesthetically pleasing, but it does the job. If you highlight areas where you think documentation deficient I can amend.

> is it possible to ensure somehow that for given input you always get the same output

I can look into the use of random.seed for this purpose

> is (-p0*p1 + p2 + x**2/p0)**(-x*(p1**p1)**x) equivalent to this LINK

The link is broken. There is no way to guarantee absolute simplification. As per the Sympy documentation, simplification is not clearly defined. They use heuristic methods that may change in the future and we view this as good enough.

> As changes were made recently, the releases of the software is now out of date

I can make a release and set the version to 0.1. I develop the software live, and I respond to feedback to fix issues as they arrive through the review, so I do not see continued development as a problem. The core functionality has been in place throughout.

> was this work funded or supported by any research councils that need mentioned

This work was not funded

> You should indicate the artists for the icon, as well as the original location that you took it from, as well as any license which is applied to it

I will try but there was limited information when I looked into this previously. I also dont want to put up the entire volunteer team who worked on the Gnome icon project.

> The algorithm of Tychonievich, 2013, seems to be a novel addition to your software, did they have enough input in this project to warrent authorship also

Dr. Tychonievich was not involved in the formation or development of this work so they are not an author. They have been cited in our paper.
