### 2019-09-07 pySRURGS/pySRURGS#6 @pySRURGS
I have added the `-deterministic` flag which runs the code in a reproducible way.
The command line tests now check that there is the results table printed.
I added a note to the readme.md for the source where I got the icon and the license is GPL v 2.0 as per the linked wikimedia article.
I will make this version release 0.1.

Closing this issue unless there are issues.
