### 2019-09-07 #1675 @pySRURGS
I have added the `-deterministic` flag and made other changes to address @SirSharpest' comments.
