### 2019-09-08 #1675 @sohrabtowfighi
After acknowledging Dr. Luther Tychonievich, I added a wayback machine archive link to the blog post where Dr. Tychonievich details the algorithm mapping integers to binary trees.
