### 2019-09-18 #1675 @mbobra
@sohrabtowfighi We're almost there! Could you please go ahead and [create a new release](https://help.github.com/en/articles/creating-releases) that incorporates all of these fixes?  And can you please [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI and then put the Zenodo badge in your README.md file? After that I think we're done 🎉
