### 2019-09-18 #1675 @mbobra
@pySRURGS Can you please make sure all the references in the paper.bib file include DOIs? Thanks!
