### 2019-09-18 #1675 @sohrabtowfighi
@mbobra I have fixed references. The work of Tychonievich has no DOI.
